[corrected] get group by id

+++ c3web-server/src/main/scala/com/ifunsoftware/c3web/service/GroupService.scala
@@ -44,7 +44,7 @@ object GroupService {
     }
   }
 
-  def getGroupById(groupId: Long): Option[Group] = {
+  def getGroupById(groupId: Int): Option[Group] = {
     groupMock find (_.id == Some(groupId))
   }
